requirements.txt
pandas
requests==2.27.1
jsonschema
pyyaml